+++ pytest.ini
@@ -8,3 +8,6 @@ markers =
     menow: the one i am working on now
     ashish: the tests ashish is working on
     managed: managed lanes
+    test_ml: .
+    geography: .
+    schema: data schema tests
