+++ tests/test_roadway.py
@@ -7,10 +7,6 @@ from network_wrangler import ProjectCard
 import time
 import numpy as np
 import pandas as pd
-from network_wrangler import offset_lat_lon
-from network_wrangler import haversine_distance
-from network_wrangler import create_unique_shape_id
-from network_wrangler import create_location_reference_from_nodes
 import networkx as nx
 from shapely.geometry import LineString
 
@@ -624,34 +620,6 @@ def test_write_model_net(request):
 
     print("--Finished:", request.node.name)
 
-
-@pytest.mark.offset
-@pytest.mark.travis
-def test_lat_lon_offset(request):
-    print("\n--Starting:", request.node.name)
-
-    in_lat_lon = [-93.0903549, 44.961085]
-    print(in_lat_lon)
-
-    new_lat_lon = offset_lat_lon(in_lat_lon)
-    print(new_lat_lon)
-
-    print("--Finished:", request.node.name)
-
-
-@pytest.mark.get_dist
-@pytest.mark.travis
-def test_get_distance_bw_lat_lon(request):
-    print("\n--Starting:", request.node.name)
-
-    start = [-93.0889873, 44.966861]
-    end = [-93.08844310000001, 44.9717832]
-    dist = haversine_distance(start, end)
-    print(dist)
-
-    print("--Finished:", request.node.name)
-
-
 @pytest.mark.roadway
 @pytest.mark.travis
 def test_network_connectivity(request):
@@ -777,18 +745,8 @@ def test_existing_managed_lane_apply(request):
     print("--Finished:", request.node.name)
 
 
-@pytest.mark.test_hash
-@pytest.mark.roadway
-def test_get_unique_shape_id(request):
-    geometry = LineString([[-93.0855338, 44.9662078], [-93.0843092, 44.9656997]])
 
-    shape_id = create_unique_shape_id(geometry)
-
-    assert shape_id == "72ceb24e2c632c02f7eae5e33ed12702"
-
-    print("--Finished:", request.node.name)
-
-
+@pytest.mark.travis
 @pytest.mark.roadway
 def test_delete_roadway_shape(request):
     print("\n--Starting:", request.node.name)
@@ -820,7 +778,7 @@ def test_delete_roadway_shape(request):
 
     print("--Finished:", request.node.name)
 
-
+@pytest.mark.travis
 @pytest.mark.roadway
 def test_create_default_geometry(request):
     print("\n--Starting:", request.node.name)
@@ -844,7 +802,7 @@ def test_create_default_geometry(request):
 
     print("--Finished:", request.node.name)
 
-
+@pytest.mark.travis
 @pytest.mark.roadway
 def test_add_roadway_shape(request):
     print("\n--Starting:", request.node.name)
@@ -876,7 +834,7 @@ def test_add_roadway_shape(request):
 
     print("--Finished:", request.node.name)
 
-
+@pytest.mark.travis
 @pytest.mark.roadway
 def test_create_ml_network_shape(request):
     print("\n--Starting:", request.node.name)
