+++ tests/test_transit_with_roadnet.py
@@ -13,8 +13,8 @@ Run just the tests labeled transit using `pytest -v -m transit`
 STPAUL_DIR = os.path.join(os.getcwd(), "examples", "stpaul")
 SCRATCH_DIR = os.path.join(os.getcwd(), "scratch")
 
-
-@pytest.mark.transit_with_roadnet
+@pytest.mark.roadway
+@pytest.mark.transit
 @pytest.mark.travis
 def test_set_roadnet(request):
     print("\n--Starting:", request.node.name)
@@ -31,7 +31,8 @@ def test_set_roadnet(request):
     print("--Finished:", request.node.name)
 
 
-@pytest.mark.transit_with_roadnet
+@pytest.mark.roadway
+@pytest.mark.transit
 @pytest.mark.travis
 # @pytest.mark.skip("")
 def test_project_card(request):
@@ -109,7 +110,8 @@ def test_project_card(request):
     print("--Finished:", request.node.name)
 
 
-@pytest.mark.transit_with_roadnet
+@pytest.mark.roadway
+@pytest.mark.transit
 @pytest.mark.travis
 # @pytest.mark.skip("need to allow for creating new stops if they don't already exist in stops.txt")
 def test_wo_existing(request):
@@ -161,7 +163,8 @@ def test_wo_existing(request):
     print("--Finished:", request.node.name)
 
 
-@pytest.mark.transit_with_roadnet
+@pytest.mark.roadway
+@pytest.mark.transit
 @pytest.mark.travis
 # @pytest.mark.skip("need to update trips and nodes")
 def test_select_transit_features_by_nodes(request):
@@ -201,7 +204,8 @@ def test_select_transit_features_by_nodes(request):
     print("--Finished:", request.node.name)
 
 
-@pytest.mark.transit_with_roadnet
+@pytest.mark.roadway
+@pytest.mark.transit
 @pytest.mark.travis
 # @pytest.mark.skip("need to update trips and nodes")
 def test_select_transit_features_by_nodes(request):
